♻️  router intent_detection.py 추출 — channel + confirmation + continuation (P0-P step 5)
변경 이유
- 3 predicate (``is_engineering_channel`` / ``detect_confirmation_signal``
  / ``should_continue_existing_thread`` / ``should_start_new_thread``)
  + ``_CONFIRMATION_KEYWORDS`` lexicon 이 router 진입점 직전에서 항상
  consult 되는데 코드는 _legacy 안에 흩어져 있었다. 한 모듈에 모아서
  signal 변경 시 변경 표면을 좁힌다.

주요 변경 사항
- 새 파일 ``engineering_channel_router/intent_detection.py`` (148 줄):
  - ``_CONFIRMATION_KEYWORDS`` 22 phrase
  - 4 predicate (channel + confirmation + continue/new-thread override)
- ``_legacy.py`` 의 해당 2 블록 (29 + 101 = 130 줄) 제거 → re-export 만.
- ``__init__.py`` facade 가 ``.intent_detection`` 에서 직접 import.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- engineering_conversation 의 더 넓은 intent classifier 와는 별개 —
  본 모듈은 router 가 진입점에서 "는 engineering 채널인가? / 어느
  thread 정책?" 를 빠르게 판단할 때 쓰는 가벼운 layer.
